Add searchUsersByUserName for the search functionality, and the data class SearchUser.
Add animation to list.

Modify fetchUsersByPhoneNumber to fetch contects. I was caching these contacts before but I think it is better to fetch again.

Add banner message to top of screen when an invite is sent. Add user type to search in order to show that each user in search list either has an invite pending etc...